+++ Blurred-ios/VideoPlaybackViewController.swift
@@ -18,6 +18,14 @@ class VideoPlaybackViewController: UIViewController {
     @IBAction func backButton(_ sender: Any) {
         self.dismiss(animated: true, completion: nil)
     }
+    fileprivate var player: AVPlayer? {
+        didSet { player?.play() }
+    }
+    deinit {
+        guard let observer = playerObserver else { return }
+        NotificationCenter.default.removeObserver(observer)
+    }
+    fileprivate var playerObserver: Any?
     let myValet = Valet.valet(with: Identifier(nonEmpty: "Id")!, accessibility: .whenUnlocked)
     let tokenValet = Valet.valet(with: Identifier(nonEmpty: "Token")!, accessibility: .whenUnlocked)
     var videoURL: URL!
@@ -32,25 +40,63 @@ class VideoPlaybackViewController: UIViewController {
 
         let playerItem = AVPlayerItem(url: videoURL as URL)
         avPlayer.replaceCurrentItem(with: playerItem)
-
+        let resetPlayer = { 
+            self.avPlayer.seek(to: CMTime.zero)
+            self.avPlayer.play()
+        }
+        playerObserver = NotificationCenter.default.addObserver(forName: NSNotification.Name.AVPlayerItemDidPlayToEndTime, object: avPlayer.currentItem, queue: nil) { notification in
+            resetPlayer()
+        }
         avPlayer.play()
-
         // Do any additional setup after loading the view.
     }
     @IBAction func doneButton(_ sender: Any) {
+        startRequest()
+    }
+    func startRequest() {
+        let myActivityIndicator = UIActivityIndicatorView(style: UIActivityIndicatorView.Style.medium)
+        myActivityIndicator.center = view.center
+        myActivityIndicator.hidesWhenStopped = true
+        myActivityIndicator.startAnimating()
+        DispatchQueue.main.async {
+            self.view.addSubview(myActivityIndicator)
+        }
+        let userId: String? = myValet.string(forKey: "Id")
+        let token: String? = tokenValet.string(forKey: "Token")
+        let Id = Int(userId!)
+        let params = ["video":["id": Id!]]
+        let headers: HTTPHeaders = [
+            "Authorization": "Bearer \(token!)",
+            "Accept": "application/json"
+        ]
+        AF.upload(
+            multipartFormData: { multipartFormData in
+                for (_, _) in params {
+                    multipartFormData.append(self.videoURL, withName: "clip" , fileName: "avatar.png", mimeType: "video/mp4")
+                    multipartFormData.append("\(Id)".data(using: String.Encoding.utf8, allowLossyConversion: false)!, withName :"Id")
+                }
+        },
+            to: "http://10.0.0.2:3000/api/v1/videouploads.json", method: .post, headers: headers)
+            .response { resp in
+                print(resp)
+
+
+        }
+        self.removeActivityIndicator(activityIndicator: myActivityIndicator)
     }
     // I want to die
     func showErrorContactingServer() {
         DispatchQueue.main.async {
             let alert = UIAlertController(title: "Error", message: "Error contacting the server. Try again later.", preferredStyle: UIAlertController.Style.alert)
 
-            // add an action (button)
+            // add an action (button)z
             alert.addAction(UIAlertAction(title: "OK", style: UIAlertAction.Style.default, handler: nil))
 
             // show the alert
             self.present(alert, animated: true, completion: nil)
         }
     }
+    //
     func showNoResponseFromServer() {
         DispatchQueue.main.async {
             let alert = UIAlertController(title: "Error", message: "No response from server. Try again later.", preferredStyle: UIAlertController.Style.alert)